pyproject.toml
[project]
name = "drf-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "dj-database-url>=3.0.1",
    "django>=5.2.5",
    "django-celery-beat>=2.8.1",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-model-utils>=5.0.0",
    "django-unfold>=0.64.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djoser>=2.3.3",
    "drf-extensions>=0.8.0",
    "flower>=2.0.1",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "redis>=6.4.0",
]

[dependency-groups]
dev = [
    "django-extensions>=4.1",
    "django-rosetta>=0.10.2",
    "django-silk>=5.4.1",
    "drf-spectacular>=0.28.0",
    "ipython>=9.4.0",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
    "black>=25.1.0",
    "isort>=6.0.1",
    "factory-boy>=3.3.3",
    "coverage>=7.10.5",
    "pydantic>=2.11.7",
    "icecream>=2.1.7",
    "werkzeug>=3.1.3",
    "pytest-cov>=6.2.1",
]
